requirements.txt
cryptography>=42.0.0
Pillow>=10.2.0
numpy>=1.26.0
scikit-image>=0.22.0
matplotlib>=3.8.0
pandas>=2.2.0
streamlit>=1.36.0
scipy>=1.11.0
gradio>=4.44.0